backend/requirements.txt
flask==3.0.3
gunicorn==23.0.0
openai>=1.0.0
python-docx
numpy
